服务列表orm映射

+++ dao/admin.go
@@ -10,7 +10,7 @@ import (
 )
 
 type Admin struct {
-	Id        int       `json:"id" gorm:"primary_key" description:"自增主键"`
+	Id        int64     `json:"id" gorm:"primary_key" description:"自增主键"`
 	UserName  string    `json:"user_name" gorm:"column:user_name" description:"管理员用户名"`
 	Salt      string    `json:"salt" gorm:"column:salt" description:"盐"`
 	Password  string    `json:"password" gorm:"column:password" description:"密码"`
